Adds oauth fix to allow refreshing without having to 'authorize' every single page
